requirements.txt
fastapi
uvicorn[standard]
gradio
langchain
langchain-community
langchain-openai
faiss-cpu
pypdf
pytesseract
pdf2image
Pillow
python-dotenv
python-multipart
tiktoken
